requirements.txt
pymongo==3.3.0
django_background_tasks==1.2.0
Django==2.1
